requirements.txt
pypdfium2>=4.30.0
Pillow>=10.0.0
